Re-initialize cropper when zoomable option changes.
This also fixes a problem where the crop data and source weren't being
reloaded when the cropper is re-initialized.

Fixes #33

https://github.com/danielthall/ember-cropperjs/issues/33

Note that in the new test case, the diff makes it look like it's
replacing a previous test, but actually there was an exact duplicate of
the last test in the file.